# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /transactions/{transactionId}:
    get:
      operationId: TransactionsController_getById
      parameters:
        - name: transactionId
          required: true
          in: path
          schema:
            type: string
        - name: include_location
          required: false
          in: query
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction_id:
                    type: string
                  location_id:
                    type: string
                  date_mutation:
                    type: string
                    nullable: true
                    format: date-time
                  nature_mutation:
                    type: string
                    nullable: true
                  valeur_fonciere:
                    type: string
                    nullable: true
                  surface_reelle_bati:
                    type: string
                    nullable: true
                  price_per_sqm_built:
                    type: string
                    nullable: true
                  type_local:
                    type: string
                    nullable: true
                  code_type_local:
                    type: string
                    nullable: true
                  id_parcelle:
                    type: string
                    nullable: true
                  code_postal:
                    type: string
                    nullable: true
                  longitude:
                    type: number
                    nullable: true
                  latitude:
                    type: number
                    nullable: true
                  source_file:
                    type: string
                    nullable: true
                  ingested_at:
                    type: string
                    format: date-time
                  code_commune:
                    type: string
                  nom_commune:
                    type: string
                    nullable: true
                required:
                  - transaction_id
                  - location_id
                  - ingested_at
        "404":
          description: Unknown transaction_id
      tags: &a2
        - transactions
  /transactions:
    get:
      operationId: TransactionsController_list
      parameters:
        - name: include_location
          required: false
          in: query
          schema:
            type: boolean
            default: true
        - name: offset
          required: false
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          required: false
          in: query
          schema:
            type: integer
            default: 200
        - name: to
          required: false
          in: query
          description: YYYY-MM-DD (inclusive)
          schema: {}
        - name: from
          required: false
          in: query
          description: YYYY-MM-DD (inclusive)
          schema: {}
        - name: type_local
          required: false
          in: query
          schema: {}
        - name: nature_mutation
          required: false
          in: query
          schema: {}
        - name: code_commune
          required: false
          in: query
          schema: {}
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    transaction_id:
                      type: string
                    location_id:
                      type: string
                    date_mutation:
                      type: string
                      nullable: true
                      format: date-time
                    nature_mutation:
                      type: string
                      nullable: true
                    valeur_fonciere:
                      type: string
                      nullable: true
                    surface_reelle_bati:
                      type: string
                      nullable: true
                    price_per_sqm_built:
                      type: string
                      nullable: true
                    type_local:
                      type: string
                      nullable: true
                    code_type_local:
                      type: string
                      nullable: true
                    id_parcelle:
                      type: string
                      nullable: true
                    code_postal:
                      type: string
                      nullable: true
                    longitude:
                      type: number
                      nullable: true
                    latitude:
                      type: number
                      nullable: true
                    source_file:
                      type: string
                      nullable: true
                    ingested_at:
                      type: string
                      format: date-time
                    code_commune:
                      type: string
                    nom_commune:
                      type: string
                      nullable: true
                  required:
                    - transaction_id
                    - location_id
                    - ingested_at
      tags: *a2
...
info:
  title: Homepedia API
  description: API on top of dbt marts from data_platform.
  version: 1.0.0
  contact: {}
...
servers: []
...